Refactor DeviantArt and icon classes into more specific namespaces
