+++ FoliCon/Modules/Extension/DialogServiceExtensions.cs
@@ -1,4 +1,5 @@
-﻿using FoliCon.Modules.IGDB;
+﻿using FoliCon.Modules.DeviantArt;
+using FoliCon.Modules.IGDB;
 using FoliCon.Modules.TMDB;
 using NLog;
 using Logger = NLog.Logger;
