+++ FoliCon/Modules/Media/ProIcon.cs
@@ -2,7 +2,7 @@
 using NLog;
 using Logger = NLog.Logger;
 
-namespace FoliCon.Modules;
+namespace FoliCon.Modules.Media;
 
 public class ProIcon
 {
