+++ FoliCon/ViewModels/MainWindowViewModel.cs
@@ -1,4 +1,5 @@
-﻿using FoliCon.Modules.Extension;
+﻿using FoliCon.Modules.DeviantArt;
+using FoliCon.Modules.Extension;
 using FoliCon.Modules.IGDB;
 using FoliCon.Modules.TMDB;
 using FoliCon.Modules.UI;
